Update NativeUtility.cs
PROPOSED change
When the R-core subkey is not present inside of HKEY_LOCAL_MACHINE, then on line 510, rCore will come back non-null, but its realKey *will* be null.  Line 511 will evaluate to false so this if block is not run.  Then on line 521 when we try to invoke OpenSubKey, we get a null reference exception because WindowsRegistryKey.OpenSubKey calls this.realKey.OpenSubKey. I do not know if this will introduce new negative side effects.